Replace array get set with square bracket operator and minor project restructuring
