+++ README.md
@@ -2,6 +2,9 @@
 
 Our own implementation of C++ std library classes and data structures to investigate impact of different optimizations of performance.
 
+Current Data Structures:
+1. Array
+
 Collaborators: Ben, Chris
 
 **Further Notes** ( TO BE REMOVED ): https://docs.google.com/document/d/10WjLhU_cxeN_wvX2Q9tp6nzZyLks1j78UNoN0ND10WI/edit
