+++ UTD/CMakeLists.txt
@@ -3,7 +3,7 @@
 #
 
 # Add source to this project's executable.
-add_executable (UTD "UTD.cpp" "UTD.h" "Array.h")
+add_executable (UTD "src/demo/Demo.cpp" "src/demo/main.cpp")
 
 if (CMAKE_VERSION VERSION_GREATER 3.12)
   set_property(TARGET UTD PROPERTY CXX_STANDARD 20)
