--- UTD/UTD.h
@@ -1,10 +1,0 @@
-﻿// UTD.h : Include file for standard system include files,
-// or project specific include files.
-
-#pragma once
-
-#include <iostream>
-
-// TODO: Reference additional headers your program requires here.
-
-#include "Array.h"
