+++ UTD/src/demo/Demo.h
@@ -1,0 +1,10 @@
+﻿#pragma once
+
+namespace utd {
+
+    void test_with_timer_class();
+
+    void test_with_time_method();
+
+    void test_with_timer_class_raii();
+}
