+++ UTD/src/demo/main.cpp
@@ -1,0 +1,11 @@
+#include "Demo.h"
+
+int main() {
+    utd::test_with_timer_class();
+
+    utd::test_with_time_method();
+
+    utd::test_with_timer_class_raii();
+
+    return 0;
+};
